+++ src/backend/src/controllers/finance.controllers.ts
@@ -157,15 +157,17 @@ export default class FinanceController {
   static async getReimbursementRequestTeamData(req: Request, res: Response, next: NextFunction) {
     try {
       const { teamId } = req.params;
-      const { startDate, endDate } = req.query;
+      const { startDate, endDate, carNumber } = req.query;
       const parsedStartDate = typeof startDate === 'string' ? new Date(startDate) : undefined;
       const parsedEndDate = typeof endDate === 'string' ? new Date(endDate) : undefined;
+      const parsedCarNumber = typeof carNumber === 'string' ? Number(carNumber) : undefined;
 
       const rrData = await FinanceServices.getReimbursementRequestTeamData(
         req.organization,
         teamId,
         parsedStartDate,
-        parsedEndDate
+        parsedEndDate,
+        parsedCarNumber
       );
       res.status(200).json(rrData);
     } catch (error: unknown) {
@@ -176,15 +178,17 @@ export default class FinanceController {
   static async getReimbursementRequestTeamTypeData(req: Request, res: Response, next: NextFunction) {
     try {
       const { teamTypeId } = req.params;
-      const { startDate, endDate } = req.query;
+      const { startDate, endDate, carNumber } = req.query;
       const parsedStartDate = typeof startDate === 'string' ? new Date(startDate) : undefined;
       const parsedEndDate = typeof endDate === 'string' ? new Date(endDate) : undefined;
+      const parsedCarNumber = typeof carNumber === 'string' ? Number(carNumber) : undefined;
 
       const rrData = await FinanceServices.getReimbursementRequestTeamTypeData(
         req.organization,
         teamTypeId,
         parsedStartDate,
-        parsedEndDate
+        parsedEndDate,
+        parsedCarNumber
       );
       res.status(200).json(rrData);
     } catch (error: unknown) {
@@ -195,15 +199,17 @@ export default class FinanceController {
   static async getSpendingBarTeamData(req: Request, res: Response, next: NextFunction) {
     try {
       const { teamId } = req.params;
-      const { startDate, endDate } = req.query;
+      const { startDate, endDate, carNumber } = req.query;
       const parsedStartDate = typeof startDate === 'string' ? new Date(startDate) : undefined;
       const parsedEndDate = typeof endDate === 'string' ? new Date(endDate) : undefined;
+      const parsedCarNumber = typeof carNumber === 'string' ? Number(carNumber) : undefined;
 
       const spendingBarData = await FinanceServices.getSpendingBarTeamData(
         req.organization,
         teamId,
         parsedStartDate,
-        parsedEndDate
+        parsedEndDate,
+        parsedCarNumber
       );
       res.status(200).json(spendingBarData);
     } catch (error: unknown) {
@@ -214,15 +220,17 @@ export default class FinanceController {
   static async getSpendingBarTeamTypeData(req: Request, res: Response, next: NextFunction) {
     try {
       const { teamTypeId } = req.params;
-      const { startDate, endDate } = req.query;
+      const { startDate, endDate, carNumber } = req.query;
       const parsedStartDate = typeof startDate === 'string' ? new Date(startDate) : undefined;
       const parsedEndDate = typeof endDate === 'string' ? new Date(endDate) : undefined;
+      const parsedCarNumber = typeof carNumber === 'string' ? Number(carNumber) : undefined;
 
       const spendingBarData = await FinanceServices.getSpendingBarTeamTypeData(
         req.organization,
         teamTypeId,
         parsedStartDate,
-        parsedEndDate
+        parsedEndDate,
+        parsedCarNumber
       );
       res.status(200).json(spendingBarData);
     } catch (error: unknown) {
@@ -232,11 +240,17 @@ export default class FinanceController {
 
   static async getAllReimbursementRequestData(req: Request, res: Response, next: NextFunction) {
     try {
-      const { startDate, endDate } = req.query;
+      const { startDate, endDate, carNumber } = req.query;
       const parsedStartDate = typeof startDate === 'string' ? new Date(startDate) : undefined;
       const parsedEndDate = typeof endDate === 'string' ? new Date(endDate) : undefined;
+      const parsedCarNumber = typeof carNumber === 'string' ? Number(carNumber) : undefined;
 
-      const rrData = await FinanceServices.getAllReimbursementRequestData(req.organization, parsedStartDate, parsedEndDate);
+      const rrData = await FinanceServices.getAllReimbursementRequestData(
+        req.organization,
+        parsedStartDate,
+        parsedEndDate,
+        parsedCarNumber
+      );
       res.status(200).json(rrData);
     } catch (error: unknown) {
       next(error);
@@ -246,15 +260,17 @@ export default class FinanceController {
   static async getReimbursementRequestCategoryData(req: Request, res: Response, next: NextFunction) {
     try {
       const { otherReasonId } = req.params;
-      const { startDate, endDate } = req.query;
+      const { startDate, endDate, carNumber } = req.query;
       const parsedStartDate = typeof startDate === 'string' ? new Date(startDate) : undefined;
       const parsedEndDate = typeof endDate === 'string' ? new Date(endDate) : undefined;
+      const parsedCarNumber = typeof carNumber === 'string' ? Number(carNumber) : undefined;
 
       const rrData = await FinanceServices.getReimbursementRequestCategoryData(
         otherReasonId,
         req.organization,
         parsedStartDate,
-        parsedEndDate
+        parsedEndDate,
+        parsedCarNumber
       );
       res.status(200).json(rrData);
     } catch (error: unknown) {
@@ -264,11 +280,18 @@ export default class FinanceController {
 
   static async getAllSpendingBarData(req: Request, res: Response, next: NextFunction) {
     try {
-      const { startDate, endDate } = req.query;
+      const { startDate, endDate, carNumber } = req.query;
+      console.log(carNumber);
       const parsedStartDate = typeof startDate === 'string' ? new Date(startDate) : undefined;
       const parsedEndDate = typeof endDate === 'string' ? new Date(endDate) : undefined;
+      const parsedCarNumber = typeof carNumber === 'string' ? Number(carNumber) : undefined;
 
-      const spendingBarData = await FinanceServices.getAllSpendingBarData(req.organization, parsedStartDate, parsedEndDate);
+      const spendingBarData = await FinanceServices.getAllSpendingBarData(
+        req.organization,
+        parsedStartDate,
+        parsedEndDate,
+        parsedCarNumber
+      );
       res.status(200).json(spendingBarData);
     } catch (error: unknown) {
       next(error);
