+++ src/backend/src/services/finance.services.ts
@@ -6,9 +6,10 @@ import {
   SpendingBarData,
   Sponsor,
   SponsorTask,
-  SponsorTier
+  SponsorTier,
+  wbsPipe
 } from 'shared';
-import { User, Organization, Sponsor_Task } from '@prisma/client';
+import { User, Organization, Sponsor_Task, Reimbursement_Status_Type } from '@prisma/client';
 import { userHasPermission } from '../utils/users.utils';
 import {
   getSponsorQueryArgs,
@@ -26,17 +27,7 @@ import {
 import prisma from '../prisma/prisma';
 import { sponsorTransformer } from '../transformers/finance.transformer';
 import sponsorTaskTransformer from '../transformers/sponsor-task.transformer';
-import {
-  getAllReimbursementRequestData,
-  getAllSpendingBarData,
-  getReimbursementRequestCategoryData,
-  getReimbursementRequestsByDivision,
-  getReimbursementRequestsByProject,
-  getReimbursementRequestsByTeam,
-  getSpendingBarCategoryData,
-  getSpendingBarDataForProjectBudgetByDivision,
-  getSpendingBarDataForProjectBudgetByTeam
-} from '../utils/finance.utils';
+import { getProjectSegmentedWhereInput, getReimbursementRequestWhereInput } from '../utils/finance.utils';
 import { notifySponsorTaskAssignee } from '../utils/slack.utils';
 
 export default class FinanceServices {
@@ -407,72 +398,770 @@ export default class FinanceServices {
     startDate?: Date,
     endDate?: Date
   ): Promise<ReimbursementRequestData> {
-    return await getReimbursementRequestsByProject(projectId, organization.organizationId, startDate, endDate);
+    const { organizationId } = organization;
+    const project = await prisma.project.findFirst({
+      where: {
+        projectId,
+        wbsElement: {
+          organizationId,
+          dateDeleted: null
+        }
+      }
+    });
+
+    if (!project) throw new NotFoundException('Project', projectId);
+
+    const reimbursementRequests = await prisma.reimbursement_Request.findMany({
+      where: {
+        reimbursementProducts: {
+          some: {
+            reimbursementProductReason: {
+              wbsElement: {
+                project: {
+                  projectId
+                }
+              }
+            }
+          }
+        },
+        reimbursementStatuses: {
+          none: {
+            type: Reimbursement_Status_Type.DENIED
+          }
+        },
+        ...getReimbursementRequestWhereInput(startDate, endDate)
+      },
+      select: {
+        reimbursementStatuses: true,
+        totalCost: true
+      }
+    });
+
+    let pendingFinance = 0;
+    let pendingLeadership = 0;
+    let submittedToSabo = 0;
+    let reimbursed = 0;
+
+    reimbursementRequests.forEach((req) => {
+      const lastStatus = req.reimbursementStatuses.at(-1)?.type;
+
+      switch (lastStatus) {
+        case Reimbursement_Status_Type.PENDING_FINANCE:
+          pendingFinance += req.totalCost;
+          break;
+        case Reimbursement_Status_Type.PENDING_LEADERSHIP_APPROVAL:
+          pendingLeadership += req.totalCost;
+          break;
+        case Reimbursement_Status_Type.SABO_SUBMITTED:
+          submittedToSabo += req.totalCost;
+          break;
+        case Reimbursement_Status_Type.REIMBURSED:
+          reimbursed += req.totalCost;
+          break;
+        default:
+          break;
+      }
+    });
+
+    pendingFinance = pendingFinance / 100;
+    pendingLeadership = pendingLeadership / 100;
+    submittedToSabo = submittedToSabo / 100;
+    reimbursed = reimbursed / 100;
+
+    const totalBalance = reimbursementRequests.reduce((acc, curr) => acc + curr.totalCost, 0) / 100;
+
+    const available = project.budget - totalBalance;
+
+    const data: ReimbursementRequestData = {
+      totalBudget: project.budget,
+      pendingFinance,
+      pendingLeadership,
+      submittedToSabo,
+      reimbursed,
+      available
+    };
+    return data;
   }
 
   static async getReimbursementRequestTeamData(
     organization: Organization,
     teamId: string,
     startDate?: Date,
-    endDate?: Date
+    endDate?: Date,
+    carNumber?: number
   ): Promise<ReimbursementRequestData> {
-    return await getReimbursementRequestsByTeam(teamId, organization.organizationId, startDate, endDate);
+    const { organizationId } = organization;
+    const team = await prisma.team.findUnique({
+      where: {
+        organizationId,
+        dateArchived: null,
+        teamId
+      },
+      include: {
+        projects: {
+          ...getProjectSegmentedWhereInput(organizationId, startDate, endDate, carNumber),
+          include: {
+            wbsElement: true
+          }
+        }
+      }
+    });
+
+    if (!team) throw new NotFoundException('Team', teamId);
+
+    const reimbursementRequests = await prisma.reimbursement_Request.findMany({
+      where: {
+        reimbursementProducts: {
+          some: {
+            reimbursementProductReason: {
+              wbsElement: {
+                project: {
+                  teams: {
+                    some: {
+                      teamId
+                    }
+                  }
+                }
+              }
+            }
+          }
+        },
+        reimbursementStatuses: {
+          none: {
+            type: Reimbursement_Status_Type.DENIED
+          }
+        },
+        ...getReimbursementRequestWhereInput(startDate, endDate, carNumber)
+      },
+      select: {
+        reimbursementStatuses: true,
+        totalCost: true
+      }
+    });
+
+    const totalBudget = team.projects.reduce((acc, curr) => acc + curr.budget, 0);
+
+    let pendingFinance = 0;
+    let pendingLeadership = 0;
+    let submittedToSabo = 0;
+    let reimbursed = 0;
+
+    reimbursementRequests.forEach((req) => {
+      const lastStatus = req.reimbursementStatuses.at(-1)?.type;
+
+      switch (lastStatus) {
+        case Reimbursement_Status_Type.PENDING_FINANCE:
+          pendingFinance += req.totalCost;
+          break;
+        case Reimbursement_Status_Type.PENDING_LEADERSHIP_APPROVAL:
+          pendingLeadership += req.totalCost;
+          break;
+        case Reimbursement_Status_Type.SABO_SUBMITTED:
+          submittedToSabo += req.totalCost;
+          break;
+        case Reimbursement_Status_Type.REIMBURSED:
+          reimbursed += req.totalCost;
+          break;
+        default:
+          break;
+      }
+    });
+
+    pendingFinance = pendingFinance / 100;
+    pendingLeadership = pendingLeadership / 100;
+    submittedToSabo = submittedToSabo / 100;
+    reimbursed = reimbursed / 100;
+
+    const totalBalance = reimbursementRequests.reduce((acc, curr) => acc + curr.totalCost, 0) / 100;
+
+    const available = totalBudget - totalBalance;
+
+    const data: ReimbursementRequestData = {
+      totalBudget,
+      pendingFinance,
+      pendingLeadership,
+      submittedToSabo,
+      reimbursed,
+      available
+    };
+    return data;
   }
 
   static async getReimbursementRequestTeamTypeData(
     organization: Organization,
     teamTypeId: string,
     startDate?: Date,
-    endDate?: Date
+    endDate?: Date,
+    carNumber?: number
   ): Promise<ReimbursementRequestData> {
-    return await getReimbursementRequestsByDivision(teamTypeId, organization.organizationId, startDate, endDate);
+    const { organizationId } = organization;
+    const division = await prisma.team_Type.findUnique({
+      where: {
+        organizationId,
+        teamTypeId
+      },
+      include: {
+        teams: {
+          where: {
+            dateArchived: null
+          },
+          include: {
+            projects: {
+              ...getProjectSegmentedWhereInput(organizationId, startDate, endDate, carNumber),
+              include: {
+                wbsElement: true
+              }
+            }
+          }
+        }
+      }
+    });
+
+    if (!division) throw new NotFoundException('Team Type', teamTypeId);
+
+    const results: ReimbursementRequestData = {
+      totalBudget: 0,
+      pendingFinance: 0,
+      pendingLeadership: 0,
+      submittedToSabo: 0,
+      reimbursed: 0,
+      available: 0
+    };
+
+    for (const team of division.teams) {
+      const data: ReimbursementRequestData = await this.getReimbursementRequestTeamData(
+        organization,
+        team.teamId,
+        startDate,
+        endDate,
+        carNumber
+      );
+
+      results.totalBudget += data.totalBudget;
+      results.pendingFinance += data.pendingFinance;
+      results.pendingLeadership += data.pendingLeadership;
+      results.submittedToSabo += data.submittedToSabo;
+      results.reimbursed += data.reimbursed;
+      results.available += data.available;
+    }
+
+    return results;
   }
 
   static async getSpendingBarTeamData(
     organization: Organization,
     teamId: string,
     startDate?: Date,
-    endDate?: Date
+    endDate?: Date,
+    carNumber?: number
   ): Promise<SpendingBarData> {
-    return await getSpendingBarDataForProjectBudgetByTeam(teamId, organization.organizationId, startDate, endDate);
+    const { organizationId } = organization;
+    const team = await prisma.team.findUnique({
+      where: {
+        organizationId,
+        dateArchived: null,
+        teamId
+      },
+      include: {
+        projects: {
+          ...getProjectSegmentedWhereInput(organizationId, startDate, endDate, carNumber),
+          include: {
+            wbsElement: true
+          }
+        }
+      }
+    });
+
+    if (!team) throw new NotFoundException('Team', teamId);
+
+    const spendingInfoPromises = team.projects.map((project) =>
+      this.getReimbursementRequestProjectData(organization, project.projectId, startDate, endDate)
+    );
+
+    const spendingInfos = await Promise.all(spendingInfoPromises);
+
+    const data: SpendingBarData = {
+      title: `${team.teamName}`,
+      data: team.projects.map((project, index) => ({
+        title: `${wbsPipe(project.wbsElement)} - ${project.wbsElement.name}`,
+        spendingInfo: spendingInfos[index]
+      }))
+    };
+    return data;
   }
 
   static async getSpendingBarTeamTypeData(
     organization: Organization,
     teamTypeId: string,
     startDate?: Date,
-    endDate?: Date
+    endDate?: Date,
+    carNumber?: number
   ): Promise<SpendingBarData[]> {
-    return await getSpendingBarDataForProjectBudgetByDivision(teamTypeId, organization.organizationId, startDate, endDate);
+    const { organizationId } = organization;
+    const division = await prisma.team_Type.findUnique({
+      where: {
+        organizationId,
+        teamTypeId
+      },
+      include: {
+        teams: {
+          where: {
+            dateArchived: null,
+            teamTypeId
+          },
+          include: {
+            projects: {
+              ...getProjectSegmentedWhereInput(organizationId, startDate, endDate, carNumber),
+              include: {
+                wbsElement: true
+              }
+            }
+          }
+        }
+      }
+    });
+
+    if (!division) throw new NotFoundException('Team Type', teamTypeId);
+
+    const teamDataPromises = division.teams.map(async (team) => {
+      const spendingInfoPromises = team.projects.map((project) =>
+        this.getReimbursementRequestProjectData(organization, project.projectId, startDate, endDate)
+      );
+      const spendingInfos = await Promise.all(spendingInfoPromises);
+
+      return {
+        title: `${team.teamName}`,
+        data: team.projects.map((project, index) => ({
+          title: `${wbsPipe(project.wbsElement)} - ${project.wbsElement.name}`,
+          spendingInfo: spendingInfos[index]
+        }))
+      };
+    });
+
+    const data: SpendingBarData[] = await Promise.all(teamDataPromises);
+
+    return data;
   }
 
   static async getAllReimbursementRequestData(
     organization: Organization,
     startDate?: Date,
-    endDate?: Date
+    endDate?: Date,
+    carNumber?: number
   ): Promise<ReimbursementRequestData[]> {
-    return await getAllReimbursementRequestData(organization.organizationId, startDate, endDate);
+    const { organizationId } = organization;
+    const cashReimbursementRequests = await prisma.reimbursement_Request.findMany({
+      where: {
+        dateDeleted: null,
+        accountCode: { organizationId },
+        indexCode: { organizationId, name: 'CASH', code: '830667' },
+        reimbursementStatuses: {
+          none: {
+            type: Reimbursement_Status_Type.DENIED
+          }
+        },
+        ...getReimbursementRequestWhereInput(startDate, endDate, carNumber)
+      },
+      select: {
+        reimbursementStatuses: true,
+        totalCost: true
+      }
+    });
+
+    const budgetReimbursementRequests = await prisma.reimbursement_Request.findMany({
+      where: {
+        dateDeleted: null,
+        accountCode: { organizationId },
+        indexCode: { organizationId, name: 'BUDGET', code: '800462' },
+        reimbursementStatuses: {
+          none: {
+            type: Reimbursement_Status_Type.DENIED
+          }
+        },
+        ...getReimbursementRequestWhereInput(startDate, endDate, carNumber)
+      },
+      select: {
+        reimbursementStatuses: true,
+        totalCost: true
+      }
+    });
+
+    const allReimbursementRequests = await prisma.reimbursement_Request.findMany({
+      where: {
+        dateDeleted: null,
+        accountCode: { organizationId },
+        indexCode: { organizationId },
+        reimbursementStatuses: {
+          none: {
+            type: Reimbursement_Status_Type.DENIED
+          }
+        },
+        ...getReimbursementRequestWhereInput(startDate, endDate, carNumber)
+      },
+      select: {
+        reimbursementStatuses: true,
+        totalCost: true
+      }
+    });
+
+    const teams = await prisma.team.findMany({
+      where: { dateArchived: null, organizationId },
+      select: {
+        projects: {
+          where: {
+            wbsElement: {
+              ...(carNumber !== undefined && { carNumber }),
+              dateCreated: {
+                ...(startDate && {
+                  gte: startDate
+                }),
+                ...(endDate && {
+                  lte: endDate
+                })
+              }
+            }
+          }
+        }
+      }
+    });
+
+    const allTotalBudget = teams.reduce((teamAcc, team) => {
+      const teamBudget = team.projects.reduce((projAcc, project) => projAcc + project.budget, 0);
+      return teamAcc + teamBudget;
+    }, 0);
+
+    const cashTotalBudget =
+      cashReimbursementRequests.reduce((reqAcc, rr) => {
+        return reqAcc + rr.totalCost;
+      }, 0) / 100;
+
+    const budgetTotalBudget =
+      budgetReimbursementRequests.reduce((reqAcc, rr) => {
+        return reqAcc + rr.totalCost;
+      }, 0) / 100;
+
+    let allPendingFinance = 0;
+    let allPendingLeadership = 0;
+    let allSubmittedToSabo = 0;
+    let allReimbursed = 0;
+
+    allReimbursementRequests.forEach((req) => {
+      const lastStatus = req.reimbursementStatuses.at(-1)?.type;
+
+      switch (lastStatus) {
+        case Reimbursement_Status_Type.PENDING_FINANCE:
+          allPendingFinance += req.totalCost;
+          break;
+        case Reimbursement_Status_Type.PENDING_LEADERSHIP_APPROVAL:
+          allPendingLeadership += req.totalCost;
+          break;
+        case Reimbursement_Status_Type.SABO_SUBMITTED:
+          allSubmittedToSabo += req.totalCost;
+          break;
+        case Reimbursement_Status_Type.REIMBURSED:
+          allReimbursed += req.totalCost;
+          break;
+        default:
+          break;
+      }
+    });
+
+    allPendingFinance = allPendingFinance / 100;
+    allPendingLeadership = allPendingLeadership / 100;
+    allSubmittedToSabo = allSubmittedToSabo / 100;
+    allReimbursed = allReimbursed / 100;
+
+    const allTotalBalance = allReimbursementRequests.reduce((acc, curr) => acc + curr.totalCost, 0) / 100;
+
+    const allAvailable = allTotalBudget - allTotalBalance;
+
+    let cashPendingFinance = 0;
+    let cashPendingLeadership = 0;
+    let cashSubmittedToSabo = 0;
+    let cashReimbursed = 0;
+
+    cashReimbursementRequests.forEach((req) => {
+      const lastStatus = req.reimbursementStatuses.at(-1)?.type;
+
+      switch (lastStatus) {
+        case Reimbursement_Status_Type.PENDING_FINANCE:
+          cashPendingFinance += req.totalCost;
+          break;
+        case Reimbursement_Status_Type.PENDING_LEADERSHIP_APPROVAL:
+          cashPendingLeadership += req.totalCost;
+          break;
+        case Reimbursement_Status_Type.SABO_SUBMITTED:
+          cashSubmittedToSabo += req.totalCost;
+          break;
+        case Reimbursement_Status_Type.REIMBURSED:
+          cashReimbursed += req.totalCost;
+          break;
+        default:
+          break;
+      }
+    });
+
+    cashPendingFinance = cashPendingFinance / 100;
+    cashPendingLeadership = cashPendingLeadership / 100;
+    cashSubmittedToSabo = cashSubmittedToSabo / 100;
+    cashReimbursed = cashReimbursed / 100;
+
+    const cashTotalBalance = cashReimbursementRequests.reduce((acc, curr) => acc + curr.totalCost, 0) / 100;
+
+    const cashAvailable = cashTotalBudget - cashTotalBalance;
+
+    let budgetPendingFinance = 0;
+    let budgetPendingLeadership = 0;
+    let budgetSubmittedToSabo = 0;
+    let budgetReimbursed = 0;
+
+    budgetReimbursementRequests.forEach((req) => {
+      const lastStatus = req.reimbursementStatuses.at(-1)?.type;
+
+      switch (lastStatus) {
+        case Reimbursement_Status_Type.PENDING_FINANCE:
+          budgetPendingFinance += req.totalCost;
+          break;
+        case Reimbursement_Status_Type.PENDING_LEADERSHIP_APPROVAL:
+          budgetPendingLeadership += req.totalCost;
+          break;
+        case Reimbursement_Status_Type.SABO_SUBMITTED:
+          budgetSubmittedToSabo += req.totalCost;
+          break;
+        case Reimbursement_Status_Type.REIMBURSED:
+          budgetReimbursed += req.totalCost;
+          break;
+        default:
+          break;
+      }
+    });
+
+    budgetPendingFinance = budgetPendingFinance / 100;
+    budgetPendingLeadership = budgetPendingLeadership / 100;
+    budgetSubmittedToSabo = budgetSubmittedToSabo / 100;
+    budgetReimbursed = budgetReimbursed / 100;
+
+    const budgetTotalBalance = budgetReimbursementRequests.reduce((acc, curr) => acc + curr.totalCost, 0) / 100;
+
+    const budgetAvailable = budgetTotalBudget - budgetTotalBalance;
+
+    const allData: ReimbursementRequestData = {
+      totalBudget: allTotalBudget,
+      pendingFinance: allPendingFinance,
+      pendingLeadership: allPendingLeadership,
+      submittedToSabo: allSubmittedToSabo,
+      reimbursed: allReimbursed,
+      available: allAvailable
+    };
+
+    const cashData: ReimbursementRequestData = {
+      totalBudget: cashTotalBudget,
+      pendingFinance: cashPendingFinance,
+      pendingLeadership: cashPendingLeadership,
+      submittedToSabo: cashSubmittedToSabo,
+      reimbursed: cashReimbursed,
+      available: cashAvailable
+    };
+
+    const budgetData: ReimbursementRequestData = {
+      totalBudget: budgetTotalBudget,
+      pendingFinance: budgetPendingFinance,
+      pendingLeadership: budgetPendingLeadership,
+      submittedToSabo: budgetSubmittedToSabo,
+      reimbursed: budgetReimbursed,
+      available: budgetAvailable
+    };
+
+    const data: ReimbursementRequestData[] = [allData, budgetData, cashData];
+
+    return data;
   }
 
   static async getReimbursementRequestCategoryData(
     otherReasonId: string,
     organization: Organization,
     startDate?: Date,
-    endDate?: Date
+    endDate?: Date,
+    carNumber?: number
   ): Promise<ReimbursementRequestData> {
-    return await getReimbursementRequestCategoryData(otherReasonId, organization.organizationId, startDate, endDate);
+    const { organizationId } = organization;
+
+    //if a car is specified but not the date ranges, use the car to determine the date ranges
+    if (carNumber !== undefined && (!startDate || !endDate)) {
+      const car = await prisma.car.findFirst({
+        where: {
+          wbsElement: {
+            carNumber
+          }
+        }
+      });
+
+      if (!car) {
+        throw new HttpException(404, `Could not find car with car number: ${carNumber}`);
+      }
+
+      const nextCar = await prisma.car.findFirst({
+        where: {
+          wbsElement: {
+            carNumber: carNumber + 1
+          }
+        }
+      });
+
+      //if no start date, set it to the creation of the selected car
+      if (!startDate) {
+        startDate = car.dateCreated;
+      }
+
+      //if no end date, set it to the creation of the next car if it exists
+      if (!endDate && nextCar) {
+        endDate = nextCar.dateCreated;
+      }
+    }
+
+    const reimbursementRequests = await prisma.reimbursement_Request.findMany({
+      where: {
+        dateDeleted: null,
+        accountCode: { organizationId },
+        reimbursementStatuses: {
+          none: {
+            type: Reimbursement_Status_Type.DENIED
+          }
+        },
+        reimbursementProducts: {
+          some: {
+            reimbursementProductReason: {
+              otherReasonId
+            }
+          }
+        },
+        ...getReimbursementRequestWhereInput(startDate, endDate)
+      },
+      select: {
+        reimbursementStatuses: true,
+        totalCost: true
+      }
+    });
+
+    const category = await prisma.reimbursement_Product_Other_Reason.findUnique({
+      where: {
+        otherReimbursementProductReasonId: otherReasonId
+      }
+    });
+
+    if (!category) throw new NotFoundException('Reimbursement Product Other Reason', otherReasonId);
+
+    const totalBudget = category.budget;
+
+    let pendingFinance = 0;
+    let pendingLeadership = 0;
+    let submittedToSabo = 0;
+    let reimbursed = 0;
+
+    reimbursementRequests.forEach((req) => {
+      const lastStatus = req.reimbursementStatuses.at(-1)?.type;
+
+      switch (lastStatus) {
+        case Reimbursement_Status_Type.PENDING_FINANCE:
+          pendingFinance += req.totalCost;
+          break;
+        case Reimbursement_Status_Type.PENDING_LEADERSHIP_APPROVAL:
+          pendingLeadership += req.totalCost;
+          break;
+        case Reimbursement_Status_Type.SABO_SUBMITTED:
+          submittedToSabo += req.totalCost;
+          break;
+        case Reimbursement_Status_Type.REIMBURSED:
+          reimbursed += req.totalCost;
+          break;
+        default:
+          break;
+      }
+    });
+
+    const totalBalance = reimbursementRequests.reduce((acc, curr) => acc + curr.totalCost, 0);
+
+    const available = totalBudget - totalBalance;
+
+    const data: ReimbursementRequestData = {
+      totalBudget,
+      pendingFinance,
+      pendingLeadership,
+      submittedToSabo,
+      reimbursed,
+      available
+    };
+
+    return data;
   }
 
   static async getAllSpendingBarData(
     organization: Organization,
     startDate?: Date,
-    endDate?: Date
+    endDate?: Date,
+    carNumber?: number
   ): Promise<SpendingBarData[]> {
-    return await getAllSpendingBarData(organization.organizationId, startDate, endDate);
+    const { organizationId } = organization;
+    const teams = await prisma.team.findMany({
+      where: {
+        organizationId,
+        dateArchived: null
+      },
+      include: {
+        projects: {
+          ...getProjectSegmentedWhereInput(organizationId, startDate, endDate, carNumber),
+          include: {
+            wbsElement: true
+          }
+        }
+      }
+    });
+
+    const teamDataPromises = teams.map(async (team) => {
+      const spendingInfoPromises = team.projects.map((project) =>
+        this.getReimbursementRequestProjectData(organization, project.projectId, startDate, endDate)
+      );
+      const spendingInfos = await Promise.all(spendingInfoPromises);
+
+      return {
+        title: `${team.teamName}`,
+        data: team.projects.map((project, index) => ({
+          title: `${wbsPipe(project.wbsElement)} - ${project.wbsElement.name}`,
+          spendingInfo: spendingInfos[index]
+        }))
+      };
+    });
+
+    const data: SpendingBarData[] = await Promise.all(teamDataPromises);
+
+    return data;
   }
 
   static async getSpendingBarCategoryData(organization: Organization): Promise<SpendingBarData> {
-    return await getSpendingBarCategoryData(organization.organizationId);
+    const { organizationId } = organization;
+    const otherReasons = await prisma.reimbursement_Product_Other_Reason.findMany({
+      where: {
+        dateDeleted: null,
+        indexCode: {
+          organizationId
+        }
+      }
+    });
+
+    const spendingInfoPromises = otherReasons.map((r) =>
+      this.getReimbursementRequestCategoryData(r.otherReimbursementProductReasonId, organization)
+    );
+    const spendingInfos = await Promise.all(spendingInfoPromises);
+
+    const data: SpendingBarData = {
+      title: `Club Categories`,
+      data: otherReasons.map((r, index) => ({
+        title: r.name,
+        spendingInfo: spendingInfos[index]
+      }))
+    };
+
+    return data;
   }
 
   /**
