+++ src/frontend/src/hooks/finance.hooks.ts
@@ -280,40 +280,47 @@ export interface ReimbursementRequestTeamDataPayload {
   teamId: string;
   startDate?: Date;
   endDate?: Date;
+  carNumber?: number;
 }
 
 export interface ReimbursementRequestDataPayload {
   startDate?: Date;
   endDate?: Date;
+  carNumber?: number;
 }
 
 export interface ReimbursementRequestCategoryDataPayload {
   otherReasonId: string;
   startDate?: Date;
   endDate?: Date;
+  carNumber?: number;
 }
 
 export interface ReimbursementRequestTeamTypeDataPayload {
   teamTypeId: string;
   startDate?: Date;
   endDate?: Date;
+  carNumber?: number;
 }
 
 export interface SpendingBarTeamDataPayload {
   teamId: string;
   startDate?: Date;
   endDate?: Date;
+  carNumber?: number;
 }
 
 export interface SpendingBarTeamTypeDataPayload {
   teamTypeId: string;
   startDate?: Date;
   endDate?: Date;
+  carNumber?: number;
 }
 
 export interface SpendingBarDataPayload {
   startDate?: Date;
   endDate?: Date;
+  carNumber?: number;
 }
 
 /**
@@ -1024,6 +1031,7 @@ export const useGetReimbursementRequestTeamData = (reimbursementRequestData: Rei
       'reimbursement-request-team-data',
       reimbursementRequestData.endDate,
       reimbursementRequestData.startDate,
+      reimbursementRequestData.carNumber,
       reimbursementRequestData.teamId
     ],
     async () => {
@@ -1038,6 +1046,7 @@ export const useGetReimbursementRequestTeamTypeData = (reimbursementRequestData:
       'reimbursement-request-team-type-data',
       reimbursementRequestData.endDate,
       reimbursementRequestData.startDate,
+      reimbursementRequestData.carNumber,
       reimbursementRequestData.teamTypeId
     ],
     async () => {
@@ -1066,6 +1075,7 @@ export const useGetReimbursementRequestCategoryData = (reimbursementRequestData:
       'reimbursement-request-category-data',
       reimbursementRequestData.endDate,
       reimbursementRequestData.startDate,
+      reimbursementRequestData.carNumber,
       reimbursementRequestData.otherReasonId
     ],
     async () => {
@@ -1085,7 +1095,13 @@ export const useGetAllReimbursementRequestData = (reimbursementRequestData: Reim
 
 export const useGetSpendingBarTeamData = (spendingBarData: SpendingBarTeamDataPayload) =>
   useQuery<SpendingBarData, Error>(
-    ['spending-bar-team-data', spendingBarData.endDate, spendingBarData.startDate, spendingBarData.teamId],
+    [
+      'spending-bar-team-data',
+      spendingBarData.endDate,
+      spendingBarData.startDate,
+      spendingBarData.carNumber,
+      spendingBarData.teamId
+    ],
     async () => {
       const { data } = await getSpendingBarTeamData(spendingBarData);
       return data;
@@ -1094,7 +1110,13 @@ export const useGetSpendingBarTeamData = (spendingBarData: SpendingBarTeamDataPa
 
 export const useGetSpendingBarTeamTypeData = (spendingBarData: SpendingBarTeamTypeDataPayload) =>
   useQuery<SpendingBarData[], Error>(
-    ['spending-bar-team-type-data', spendingBarData.endDate, spendingBarData.startDate, spendingBarData.teamTypeId],
+    [
+      'spending-bar-team-type-data',
+      spendingBarData.endDate,
+      spendingBarData.startDate,
+      spendingBarData.carNumber,
+      spendingBarData.teamTypeId
+    ],
     async () => {
       const { data } = await getSpendingBarTeamTypeData(spendingBarData);
       return data;
@@ -1103,7 +1125,7 @@ export const useGetSpendingBarTeamTypeData = (spendingBarData: SpendingBarTeamTy
 
 export const useGetSpendingBarCategoryData = (spendingBarData: SpendingBarDataPayload) =>
   useQuery<SpendingBarData, Error>(
-    ['spending-bar-category-data', spendingBarData.endDate, spendingBarData.startDate],
+    ['spending-bar-category-data', spendingBarData.endDate, spendingBarData.startDate, spendingBarData.carNumber],
     async () => {
       const { data } = await getSpendingBarCategoryData(spendingBarData);
       return data;
@@ -1111,10 +1133,13 @@ export const useGetSpendingBarCategoryData = (spendingBarData: SpendingBarDataPa
   );
 
 export const useGetAllSpendingBarData = (spendingBarData: SpendingBarDataPayload) =>
-  useQuery<SpendingBarData[], Error>(['spending-bar-data', spendingBarData.endDate, spendingBarData.startDate], async () => {
-    const { data } = await getAllSpendingBarData(spendingBarData);
-    return data;
-  });
+  useQuery<SpendingBarData[], Error>(
+    ['spending-bar-data', spendingBarData.endDate, spendingBarData.startDate, spendingBarData.carNumber],
+    async () => {
+      const { data } = await getAllSpendingBarData(spendingBarData);
+      return data;
+    }
+  );
 
 /**
  * Custom react hook to get all sponsor tiers
