+++ src/frontend/src/pages/FinancePage/FinanceDashboard/FinanceDashboardAllView.tsx
@@ -9,10 +9,11 @@ import AdminBalance from './AdminBalance';
 interface FinanceDashboardAllViewProps {
   startDate?: Date;
   endDate?: Date;
+  carNumber?: number;
 }
 
-const FinanceDashboardAllView: React.FC<FinanceDashboardAllViewProps> = ({ startDate, endDate }) => {
-  const payload = { startDate, endDate };
+const FinanceDashboardAllView: React.FC<FinanceDashboardAllViewProps> = ({ startDate, endDate, carNumber }) => {
+  const payload = { startDate, endDate, carNumber };
   // this hook returns the all data then budget data then cash data
   const {
     data: allRRData,
