+++ src/frontend/src/pages/FinancePage/FinanceDashboard/FinanceDashboardCategoriesView.tsx
@@ -8,23 +8,24 @@ import AdminBalance from './AdminBalance';
 interface FinanceDashboardCategoryViewProps {
   startDate?: Date;
   endDate?: Date;
+  carNumber?: number;
 }
 
-const FinanceDashboardCategoriesView: React.FC<FinanceDashboardCategoryViewProps> = ({ startDate, endDate }) => {
+const FinanceDashboardCategoriesView: React.FC<FinanceDashboardCategoryViewProps> = ({ startDate, endDate, carNumber }) => {
   // this hook returns the all data then budget data then cash data
   const {
     data: rrData,
     isLoading: rrDataIsLoading,
     isError: rrDataIsError,
     error: rrDataError
-  } = useGetAllReimbursementRequestData({ startDate, endDate });
+  } = useGetAllReimbursementRequestData({ startDate, endDate, carNumber });
 
   const {
     data: spendingBarData,
     isLoading: spendingBarDataIsLoading,
     isError: spendingBarDataIsError,
     error: spendingBarDataError
-  } = useGetSpendingBarCategoryData({ startDate, endDate });
+  } = useGetSpendingBarCategoryData({ startDate, endDate, carNumber });
 
   if (rrDataIsError) {
     return <ErrorPage error={rrDataError} />;
